Changed sweep to registry. update readme
